Revert "use snapshots to fix tests"
This reverts commit 01bc412864ef0132866d5c039395cf3162aabfe0.
