--- packages/core/src/__tests__/data/builder/regen-fragment.tsx
@@ -1,10 +1,0 @@
-export default function MyComponent(props) {
-  return (
-    <>
-      Hello world
-      <>
-        <Fragment>Hi</Fragment>
-      </>
-    </>
-  );
-}
