--- packages/core/src/__tests__/data/builder/regen-span-text.tsx
@@ -1,18 +1,0 @@
-export default function MyComponent(props) {
-  return (
-    <div
-      css={{
-        display: 'block',
-      }}
-    >
-      Hi there
-      <span
-        css={{
-          color: 'red',
-        }}
-      >
-        Hello world
-      </span>
-    </div>
-  );
-}
